refactor(category combo select): move into its own dedicated folder
