+++ src/components/metadataSelects/CategoryComboSelect/index.ts
@@ -1,0 +1,1 @@
+export { CategoryComboSelect } from './CategoryComboSelect'
