+++ src/components/metadataSelects/CategoryComboSelect/types.ts
@@ -1,4 +1,4 @@
-import { CategoryCombo, GistCollectionResponse } from '../../types/generated'
+import { CategoryCombo, GistCollectionResponse } from '../../../types/generated'
 
 const filterFields = ['id', 'name'] as const //(name is translated by default in /gist)
 export type FilteredCategoryCombo = Pick<
