+++ src/components/metadataSelects/index.ts
@@ -1,1 +1,1 @@
-export { CategoryComboSelect } from './CategoryComboSelect'
+export * from './CategoryComboSelect'
